+++ tests/integrationtests/HttpTestUtil.h
@@ -28,9 +28,9 @@ public:
     static Poco::Path createDatabasePath(const std::string parentPath);
     static std::string createCachedResponsedBodyFilePath(const std::string& cachePath, Request::HttpMethod httpMethod,
             const std::string& url);
-    static std::string makeUrl(const std::string& scheme, const std::string& host, unsigned int port,
+    static std::string makeUrl(const std::string& scheme, const std::string& host, unsigned short port,
             const std::string& path);
-    static std::string makeUrl(const std::string& scheme, const std::string& host, unsigned int port,
+    static std::string makeUrl(const std::string& scheme, const std::string& host, unsigned short port,
             const std::string& path, const std::string& query);
     static size_t readAllData(ResponseBodyStream::Ptr pResponseBodyStream, Poco::Buffer<char>& buffer);
     static void loadDefaultCertData();
