+++ tests/testutil/HttpTestServer.h
@@ -16,7 +16,7 @@ public:
     virtual ~HttpTestServer();
 
 protected:
-    virtual Poco::Net::ServerSocket* newSocket(unsigned int port);
+    virtual Poco::Net::ServerSocket* newSocket(unsigned short port);
 
 };
